Fix the sky build instructions
Sky can't really build on its own.  We were previously
using the 'sky' build target for this but it didn't
fully anticipate all of our dependencies.  The most
recent break was when James removed /mojo/shell
from mojo.

Instead of sky listing out each and every service
which it depends on, we should just get used to
building root like other mojo developers.

Thankfully there is already a helpful tool
for this called mojob.py.  I've updated our
readme to reflect this, and removed the mojo
dependency on the sky build group.

The purpose of the sky build group is to expose
all of the sky build targets up to the root level
BUILD.gn (which only depends on /sky), not to
be an independently buildable target.

R=jamesr@chromium.org, jamesr
BUG=

Review URL: https://codereview.chromium.org/764023007